+++ packages/eip-5792-middleware/src/constants.ts
@@ -11,12 +11,23 @@ export enum MessageType {
   SendTransaction = 'eth_sendTransaction',
 }
 
+export enum SupportedCapabilities {
+  AuxiliaryFunds = 'auxiliaryFunds',
+}
+
 // To be moved to @metamask/rpc-errors in future.
 export enum EIP5792ErrorCode {
   UnsupportedNonOptionalCapability = 5700,
   UnsupportedChainId = 5710,
   UnknownBundleId = 5730,
   RejectedUpgrade = 5750,
+}
+
+// To be moved to @metamask/rpc-errors in future.
+export enum EIP7682ErrorCode {
+  UnsupportedAsset = 5771,
+  UnsupportedChain = 5772,
+  MalformedRequiredAssets = 5773,
 }
 
 // wallet_getCallStatus