+++ packages/notification-services-controller/src/NotificationServicesController/NotificationServicesController.ts
@@ -273,6 +273,7 @@ type FeatureAnnouncementEnv = {
   spaceId: string;
   accessToken: string;
   platform: 'extension' | 'mobile';
+  platformVersion?: string;
 };
 
 /**